+++ Lib/DatabaseWrapper.h
@@ -14,16 +14,16 @@ public:
     DatabaseWrapper(std::shared_ptr<VectorFile> vectorFile);
     virtual ~DatabaseWrapper();
 
-    int SaveAreaOfInterest(const Area& areaOfInterest);
-    int SaveTile(int parentAreaId, std::string uuid, const Area& tileArea);
-    int SaveTileFile(int tileId, std::string filename, std::string layerName, int year);
-    int SaveBuilding(int tileId, const Feature& buildingFeature);
+    long long SaveAreaOfInterest(const Area& areaOfInterest);
+    long long SaveTile(int parentAreaId, std::string uuid, const Area& tileArea);
+    long long SaveTileFile(int tileId, std::string filename, std::string layerName, int year);
+    long long SaveBuilding(int tileId, const Feature& buildingFeature);
 
     static shared_ptr<DatabaseWrapper> DatabaseWrapperFor(std::string vectorFilename);
 
 private:
-    int SaveFeature(std::string tableName, std::function<void(Feature&)> saveFunction);
-    int SaveFeature(std::string tableName, const Feature& feature);
+    long long SaveFeature(std::string tableName, std::function<void(Feature&)> saveFunction);
+    long long SaveFeature(std::string tableName, const Feature& feature);
 
 private:
     std::shared_ptr<VectorFile> _vectorFile;
